requirements.txt
Django==4.1.13
djangorestframework==3.14.0
django-cors-headers==4.3.1
pymongo==3.12.3
djongo==1.3.6
sqlparse==0.2.4
python-dotenv
celery
redis
djangorestframework-simplejwt
uvicorn
gunicorn
requests
beautifulsoup4
pandas
openpyxl
